// src/types/etherscan.types.ts
export interface EtherscanTransaction {
  blockNumber: string
  timeStamp: string
  hash: string
  nonce: string
  blockHash: string
  transactionIndex: string
  from: string
  to: string
  value: string
  gas: string
  gasPrice: string
  isError: string
  txreceipt_status: string
  input: string
  contractAddress: string
  cumulativeGasUsed: string
  gasUsed: string
  confirmations: string
  methodId: string
  functionName: string
}

export interface EtherScanTransactionSimplified {
  blockNumber: string
  from: string
  to: string
  timeStamp: string
  hash: string
  value: string
}